refactor: create where handler for handlebars
