+++ build.gradle
@@ -146,7 +146,7 @@ artifacts {
 test {
     useJUnitPlatform()
     filter {
-        includeTestsMatching 'com.joutvhu.dynamic.jpa.*'
+        includeTestsMatching 'com.joutvhu.dynamic.*'
     }
 }
